Move blueprint functions from generator-jhipster-private to blueprint.
